Feature: Add syslog logger
This implements RFC5424 version of the protocol.

Doesn't use https://github.com/arcao/Syslog since the protocol itself
is trivial and most of the libraries functionality is not needed here.
The library also doesn't support setting the PROCID field, which is set
to a random id to indicate a reboot here.

Adds UI for syslog configuration to network admin view.

+++ include/Configuration.h
@@ -81,7 +81,7 @@ struct CONFIG_T {
     struct {
         bool Enabled;
         char Hostname[SYSLOG_MAX_HOSTNAME_STRLEN + 1];
-        uint32_t Port;
+        uint16_t Port;
     } Syslog;
 
     struct {
